feat: adding code splitting example. Ref sn
